requirement.txt
streamlit
huggingface_hub
Pillow